Lots of code cleanup
