+++ diffusion/train.py
@@ -18,6 +18,38 @@ from composer.core import Precision
 from composer.loggers import LoggerDestination
 from composer.utils import dist, reproducibility
 from omegaconf import DictConfig, OmegaConf
+from torch.optim import Optimizer
+
+
+def make_autoencoder_optimizer(config: DictConfig, model: ComposerModel) -> Optimizer:
+    """Configures the optimizer for use with an autoencoder + discriminator loss."""
+    print('Configuring opimizer for autoencoder+discriminator')
+    assert isinstance(model.autoencoder_loss, nn.Module)
+
+    # Configure optimizer settings for the autoencoder
+    if hasattr(config, 'autoencoder_optimizer'):
+        autoencoder_param_dict = {k: v for k, v in config.autoencoder_optimizer.items()}
+    else:
+        autoencoder_param_dict = {k: v for k, v in config.optimizer.items()}
+
+    assert hasattr(model, 'model') and isinstance(model.model, nn.Module)
+    if model.autoencoder_loss.learn_log_var:
+        autoencoder_param_dict['params'] = chain(model.model.parameters(), [model.autoencoder_loss.log_var])
+    else:
+        autoencoder_param_dict['params'] = model.model.parameters()
+
+    # Configure optimizer settings for the discriminator
+    assert hasattr(model.autoencoder_loss, 'discriminator')
+    assert isinstance(model.autoencoder_loss.discriminator, nn.Module)
+    if hasattr(config, 'discriminator_optimizer'):
+        discriminator_param_dict = {k: v for k, v in config.discriminator_optimizer.items()}
+    else:
+        discriminator_param_dict = {k: v for k, v in config.optimizer.items()}
+    discriminator_param_dict['params'] = model.autoencoder_loss.discriminator.parameters()
+
+    params = [autoencoder_param_dict, discriminator_param_dict]
+    optimizer = hydra.utils.instantiate(config.optimizer, params)
+    return optimizer
 
 
 def train(config: DictConfig) -> None:
@@ -34,32 +66,7 @@ def train(config: DictConfig) -> None:
 
     # Check if this is training an autoencoder. If so, the optimizer needs different param groups
     if hasattr(model, 'autoencoder_loss'):
-        print('Configuring opimizer for autoencoder+discriminator')
-        assert isinstance(model.autoencoder_loss, nn.Module)
-
-        # Configure optimizer settings for the autoencoder
-        if hasattr(config, 'autoencoder_optimizer'):
-            autoencoder_param_dict = {k: v for k, v in config.autoencoder_optimizer.items()}
-        else:
-            autoencoder_param_dict = {k: v for k, v in config.optimizer.items()}
-
-        assert hasattr(model, 'model') and isinstance(model.model, nn.Module)
-        if model.autoencoder_loss.learn_log_var:
-            autoencoder_param_dict['params'] = chain(model.model.parameters(), [model.autoencoder_loss.log_var])
-        else:
-            autoencoder_param_dict['params'] = model.model.parameters()
-
-        # Configure optimizer settings for the discriminator
-        assert hasattr(model.autoencoder_loss, 'discriminator')
-        assert isinstance(model.autoencoder_loss.discriminator, nn.Module)
-        if hasattr(config, 'discriminator_optimizer'):
-            discriminator_param_dict = {k: v for k, v in config.discriminator_optimizer.items()}
-        else:
-            discriminator_param_dict = {k: v for k, v in config.optimizer.items()}
-        discriminator_param_dict['params'] = model.autoencoder_loss.discriminator.parameters()
-
-        params = [autoencoder_param_dict, discriminator_param_dict]
-        optimizer = hydra.utils.instantiate(config.optimizer, params)
+        optimizer = make_autoencoder_optimizer(config, model)
     else:
         optimizer = hydra.utils.instantiate(config.optimizer, params=model.parameters())
